Revert: T: Introduce @ProjectDescriptor annotation
It allows to set project descriptor (like
WithStdRustProjectDescriptor) for a specific
test method. Previously it was possible to set
project descriptor for entire test class only. (baf1953)
