+++ src/test/kotlin/org/rust/cargo/runconfig/filters/RsBacktraceFilterTest.kt
@@ -7,16 +7,17 @@ package org.rust.cargo.runconfig.filters
 
 import com.intellij.openapi.util.SystemInfo
 import org.rust.cargo.project.model.cargoProjects
+import org.rust.lang.ProjectDescriptor
+import org.rust.lang.WithStdlibRustProjectDescriptor
 
 /**
  * Tests for RustBacktraceFilter
  */
+@ProjectDescriptor(WithStdlibRustProjectDescriptor::class)
 class RsBacktraceFilterTest : HighlightFilterTestBase() {
     private val filter: RsBacktraceFilter
         get() =
             RsBacktraceFilter(project, projectDir, project.cargoProjects.allProjects.single().workspace)
-
-    override fun getProjectDescriptor() = WithStdlibRustProjectDescriptor
 
     fun `test rustc source code link`() =
         checkHighlights(filter,
