+++ src/test/kotlin/org/rust/ide/annotator/RsStdlibErrorAnnotatorTest.kt
@@ -5,16 +5,11 @@
 
 package org.rust.ide.annotator
 
-import org.rust.lang.RsTestBase
+import org.rust.lang.ProjectDescriptor
+import org.rust.lang.WithStdlibAndDependencyRustProjectDescriptor
 
+@ProjectDescriptor(WithStdlibAndDependencyRustProjectDescriptor::class)
 class RsStdlibErrorAnnotatorTest : RsAnnotationTestBase() {
-    override fun getProjectDescriptor() = RsTestBase.WithStdlibAndDependencyRustProjectDescriptor
-
-    override fun setUp() {
-        super.setUp()
-        projectDescriptor.setUp(myFixture)
-    }
-
     fun `test E0428 respects crate aliases`() = checkErrors("""
         extern crate libc as libc_alias;
         mod libc {}
