+++ src/test/kotlin/org/rust/ide/annotator/fixes/ConvertToOwnedTyFixTest.kt
@@ -7,10 +7,11 @@ package org.rust.ide.annotator.fixes
 
 import org.rust.ide.inspections.RsInspectionsTestBase
 import org.rust.ide.inspections.RsTypeCheckInspection
+import org.rust.lang.ProjectDescriptor
+import org.rust.lang.WithStdlibRustProjectDescriptor
 
+@ProjectDescriptor(WithStdlibRustProjectDescriptor::class)
 class ConvertToOwnedTyFixTest : RsInspectionsTestBase(RsTypeCheckInspection()) {
-    override fun getProjectDescriptor() = WithStdlibRustProjectDescriptor
-
     fun `test B is Owned type of A`() = checkFixByText("Convert to B using `ToOwned` trait","""
         use std::borrow::Borrow;
