+++ src/test/kotlin/org/rust/ide/docs/RsExternalDocUrlTest.kt
@@ -5,12 +5,11 @@
 
 package org.rust.ide.docs
 
-import com.intellij.testFramework.LightProjectDescriptor
+import org.rust.lang.ProjectDescriptor
+import org.rust.lang.WithStdlibAndDependencyRustProjectDescriptor
 
+@ProjectDescriptor(WithStdlibAndDependencyRustProjectDescriptor::class)
 class RsExternalDocUrlTest : RsDocumentationProviderTest() {
-
-    override fun getProjectDescriptor(): LightProjectDescriptor = WithStdlibAndDependencyRustProjectDescriptor
-
     fun `test not stdlib item`() = doUrlTestByFileTree("""
         //- dep-lib/lib.rs
         pub struct Foo;
