+++ src/test/kotlin/org/rust/ide/inspections/RsLintLevelTest.kt
@@ -8,7 +8,7 @@ package org.rust.ide.inspections
 /**
  * Tests for lint level detection.
  */
-class RsLintLevelStructTest : RsInspectionsTestBase(RsStructNamingInspection(), true) {
+class RsLintLevelStructTest : RsInspectionsTestBase(RsStructNamingInspection()) {
 
     fun testDirrectAllow() = checkByText("""
         #[allow(non_camel_case_types)]
@@ -101,7 +101,7 @@ class RsLintLevelStructTest : RsInspectionsTestBase(RsStructNamingInspection(), 
     """)
 }
 
-class RsLintLevelFieldTest : RsInspectionsTestBase(RsFieldNamingInspection(), true) {
+class RsLintLevelFieldTest : RsInspectionsTestBase(RsFieldNamingInspection()) {
 
     fun testParentAllow() = checkByText("""
         #[allow(non_snake_case)]
