+++ src/test/kotlin/org/rust/lang/core/completion/RsExternCrateCompletionTest.kt
@@ -5,14 +5,11 @@
 
 package org.rust.lang.core.completion
 
+import org.rust.lang.ProjectDescriptor
+import org.rust.lang.WithStdlibAndDependencyRustProjectDescriptor
+
+@ProjectDescriptor(WithStdlibAndDependencyRustProjectDescriptor::class)
 class RsExternCrateCompletionTest : RsCompletionTestBase() {
-    override fun getProjectDescriptor() = WithStdlibAndDependencyRustProjectDescriptor
-
-    override fun setUp() {
-        super.setUp()
-        projectDescriptor.setUp(myFixture)
-    }
-
     fun `test extern crate`() = doSingleCompletion(
         "extern crate dep_l/*caret*/",
         "extern crate dep_lib_target/*caret*/"
